backend/HiMem/requirements.txt
transformers>=4.30.1
fire
tokenizers>=0.13.1
gradio
langchain
langchain-openai
langchain_community
segeval
nltk
scikit-learn
qdrant-client
sentence-transformers
posthog
loguru
pendulum
parsedatetime
pyyaml
langchain-neo4j
rank-bm25
opensearch-py==3.1.0
faiss-cpu
#Benchmark
rouge-score
bert-score
#Optional
psycopg[binary]
psycopg[pool]